+++ src/index.html
@@ -24,7 +24,7 @@
             <h2>Add your score</h2>
             <div class="add-score-input-container">
                 <input class="add-score-input-text" type="text" id="name" name="name" placeholder="Your name" required minlength="2" maxlength="100" size="20">
-                <input class="add-score-input-text" type="number" id="score" name="score" placeholder="Your score" required size="20">
+                <input class="add-score-input-text" type="number" id="score" name="score" placeholder="Your score" required>
             </div>
             <button id="add-score-button" type="button">Submit</button>
         </form>
